requirements.txt
numpy==1.19.5
Pillow==8.3.1
streamlit==0.86.0
streamlit-drawable-canvas==0.8.0
tensorflow==2.5.0